+++ eslint.config.js
@@ -1,6 +1,7 @@
 import js from '@eslint/js';
 import tseslint from '@typescript-eslint/eslint-plugin';
 import tsparser from '@typescript-eslint/parser';
+import globals from 'globals';
 
 export default [
   js.configs.recommended,
@@ -12,6 +13,7 @@ export default [
         ecmaVersion: 'latest',
         sourceType: 'module',
       },
+      globals: globals.browser,
     },
     plugins: {
       '@typescript-eslint': tseslint,